get parameter help
