+++ src/libraries/Microsoft.PowerFx.Core/Public/Reflection/FunctionInfoSignature.cs
@@ -13,99 +13,6 @@ using Microsoft.PowerFx.Core.Types;
 namespace Microsoft.PowerFx
 {
     /// <summary>
-    /// Information about a built-in function.
-    /// </summary>
-    [ThreadSafeImmutable]
-    [DebuggerDisplay("{Name}")]
-    public class FunctionInfo
-    {
-        internal readonly TexlFunction _fnc;
-
-        internal FunctionInfo(TexlFunction fnc)
-        {
-            _fnc = fnc ?? throw new ArgumentNullException(nameof(fnc));
-        }
-
-        /// <summary>
-        /// Name of the function.
-        /// </summary>
-        public string Name => _fnc.Name;
-
-        /// <summary>
-        /// Minimal arity of the function.
-        /// </summary>
-        public int MinArity => _fnc.MinArity;
-
-        /// <summary>
-        /// Maximal arity of the function.
-        /// </summary>
-        public int MaxArity => _fnc.MaxArity;
-
-        /// <summary>
-        /// Get a short description of the function.  
-        /// </summary>
-        [Obsolete("Use GetDescription() which takes a locale")]
-        public string Description => _fnc.Description;
-
-        /// <summary>
-        /// Get a short description of the function.
-        /// </summary>
-        /// <param name="culture"></param>
-        /// <returns></returns>
-        public string GetDescription(CultureInfo culture = null)
-        {
-            culture ??= CultureInfo.InvariantCulture;
-            return _fnc.GetDescription(culture.Name);
-        }
-
-        /// <summary>
-        /// An optional URL for more help on this function. 
-        /// </summary>
-        public string HelpLink => _fnc.HelpLink;
-
-        /// <summary>
-        /// Get parameter names for the specific overload.
-        /// </summary>
-        /// <returns></returns>
-        public IEnumerable<FunctionInfoSignature> Signatures 
-        {
-            get
-            {
-                var sigs = _fnc.GetSignatures();
-
-                foreach (var paramList in sigs)
-                {
-                    yield return new FunctionInfoSignature(this, paramList);
-                }
-            }
-        }
-
-        public IEnumerable<string> GetCategoryNames(CultureInfo culture = null)
-        {
-            culture ??= CultureInfo.InvariantCulture;
-            foreach (FunctionCategories category in Enum.GetValues(typeof(FunctionCategories)))
-            {
-                if ((_fnc.FunctionCategoriesMask & category) != 0)
-                {
-                    yield return category.GetLocalizedName(culture);
-                }
-            }
-        }
-    }
-
-    // ParameterInfo may have been a better name, but that conflicts with a type
-    // in Microsoft.PowerFx.Core.Functions.TransportSchemas. 
-    [DebuggerDisplay("{Name}: {Description}")]
-    public class ParameterInfoSignature
-    {
-        public string Name { get; init;  }
-
-        public string Description { get; init; }
-
-        // More fields, like optional? default value?, etc.
-    }
-
-    /// <summary>
     /// Represent a possible signature for a <see cref="FunctionInfo"/>.
     /// </summary>
     [DebuggerDisplay("{DebugToString()}")]
@@ -146,13 +53,13 @@ namespace Microsoft.PowerFx
             foreach (var p in _paramNames)
             {
                 string unalterableName = p(localeName);
-                
+
                 string invariantParamName = p("en-US");
 
                 // We should allow passing in culture to get the help text. 
                 // https://github.com/microsoft/Power-Fx/issues/2216
                 _parent._fnc.TryGetParamDescription(invariantParamName, out var description);
-                
+
                 result.Add(new ParameterInfoSignature
                 {
                     Name = unalterableName,
@@ -182,7 +89,7 @@ namespace Microsoft.PowerFx
 
             return sb.ToString();
         }
-        
+
         internal FunctionInfoSignature(FunctionInfo parent, TexlStrings.StringGetter[] paramNames)
         {
             _parent = parent;
